Brain-AI-Framework/platform/saas/brain_ai_saas/backend/requirements.txt
# Brain AI SaaS Backend Dependencies

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client & External APIs
httpx==0.25.2
requests==2.31.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Vector Database
pinecone-client==2.2.4
sentence-transformers==2.2.2

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1